refactor(rst): unify docinfo metadata normalization
The JS fallback parser (rst.ts) and the Python renderer bridge
(rst-renderer.ts) each carried their own parseBoolean/parseDate/parseSort/
parseType and field switch, so the two rST paths could normalize the same
document differently. src/lib/rst-metadata.ts now owns the RstMetadata
shape, the per-field validators, and RstParseError; both paths are thin
adapters over it. A new integration contract test pins Python/JS metadata
parity on a real fixture and documents the remaining heading-extraction
divergence. RST_RENDERER_DISK_CACHE_VERSION bumped 3→4 as insurance
against any cached-output drift.
